Create paintJob.Java
This code will calculate the needed amounts of buckets of paint that will be needed to paint a wall, based on the given information. In the first method, for parameters worth of information is known. In the second overloaded method, three parameters are known and in the last, two are known.