Create wallArea.java
The class has two constructors. The first constructor does not have any parameters (no-arg constructor). The second constructor has parameters width and height of type double and it initializes the fields. In case the width is less than 0 it sets the width field value to 0, in case the height parameter is less than 0 it sets the height field value to 0.